Added more images to resources for testing, added a progress bar, added an alert when the reset options button is pressed to confirm you really want to reset the options.
useTranslation property of the Stitcher was removed, because using a translation for mismatches is always performed.

progressMax is now set in the Stitcher init, and not in stitchImages, now set to 11 since I added more points that progress is displayed. The bottleneck in processing is key point matching, not computing the homography.

useLastMinSquaredDistancePercent now defaults to YES because it seems it works better than not checking the threshold.