Added NSStringExtensions, for extracting substrings delimited by border strings.
Add a UIPickerView from which you can choose which resource images to stitch. This involved putting a transparent button over the image view you can click to display the image picker, that fades in at the bottom, over the text view.

DoubleShot view controller is now the dataSource and delegate for the picker.

Image names, stored in the ivar imageNames, is the data source constructed by iterating through all image resources, and finding those that begin with “left_”